utils/requirements.txt
pandas~=2.2.3
numpy~=2.2.0
networkx~=3.4.2
scipy~=1.14.1
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
fireducks~=1.1.4
requests~=2.32.3
hdbscan~=0.8.40
scikit-learn~=1.6.0
pyproj~=3.7.0
neo4j~=5.27.0
psycopg2-binary~=2.9.10
geopandas==0.14.0
pyogrio==0.10.0
kagglehub==0.3.6
shapely==2.0.6